email display on external presenter api (#620)
* email display on external presenter api

* app package integration add email on app user serialization

* app container settings page

* fix app packages spec
