+++ app/javascript/src/layout/sidebar.tsx
@@ -5,8 +5,6 @@ import Tooltip from 'rc-tooltip';
 import icon from '../images/favicon.png';
 import {
   MoreIcon,
-  WebhooksIcon,
-  ApiIcon,
   DashboardIcon,
   PlatformIcon,
   ConversationChatIcon,
@@ -24,11 +22,6 @@ import {
   ArticlesIcon,
   CollectionsIcon,
   ChatIcon,
-  BillingIcon,
-  IntegrationsIcon,
-  TeamIcon,
-  MessengerIcon,
-  AppSettingsIcon,
   DarkModeIcon,
   LightModeIcon,
   ChartsIcons,
@@ -58,7 +51,7 @@ import { signout } from '@chaskiq/store/src/actions/auth';
 import SwitchControl from '@chaskiq/components/src/components/Switch';
 
 import { allowedAccessTo } from '@chaskiq/components/src/components/AccessDenied';
-
+import app_settings_items from './settingsItems';
 declare global {
   interface Window {
     location: Location;
@@ -396,68 +389,7 @@ function Sidebar({
         </svg>
       ),
       url: `/apps/${app.key}/settings`,
-      children: [
-        {
-          id: 'App Settings',
-          label: I18n.t('navigator.childs.app_settings'),
-          icon: <AppSettingsIcon />,
-          url: `/apps/${app.key}/settings`,
-          active: isActivePage('app_settings'),
-          allowed: allowedAccessTo(app, 'app_settings'),
-        },
-
-        {
-          id: 'Messenger',
-          label: I18n.t('navigator.childs.messenger_settings'),
-          icon: <MessengerIcon />,
-          url: `/apps/${app.key}/messenger`,
-          active: isActivePage('messenger'),
-          allowed: allowedAccessTo(app, 'app_settings'),
-        },
-
-        {
-          id: 'Team',
-          label: I18n.t('navigator.childs.team'),
-          icon: <TeamIcon />,
-          url: `/apps/${app.key}/team`,
-          active: isActivePage('team'),
-          allowed: allowedAccessTo(app, 'team'),
-        },
-        {
-          id: 'Integrations',
-          label: I18n.t('navigator.childs.integrations'),
-          icon: <IntegrationsIcon />,
-          url: `/apps/${app.key}/integrations`,
-          active: isActivePage('integrations'),
-          allowed: allowedAccessTo(app, 'app_packages'),
-        },
-        {
-          id: 'Webhooks',
-          label: I18n.t('navigator.childs.webhooks'),
-          icon: <WebhooksIcon />,
-          url: `/apps/${app.key}/webhooks`,
-          active: isActivePage('webhooks'),
-          allowed: allowedAccessTo(app, 'outgoing_webhooks'),
-        },
-        {
-          id: 'Api access',
-          label: I18n.t('navigator.childs.api_access'),
-          icon: <ApiIcon />,
-          url: `/apps/${app.key}/oauth_applications`,
-          active: isActivePage('oauth_applications'),
-          allowed: allowedAccessTo(app, 'oauth_applications'),
-        },
-        {
-          id: 'Billing',
-          icon: <BillingIcon />,
-          hidden: !app.subscriptionsEnabled,
-          label: I18n.t('navigator.childs.billing'),
-          url: `/apps/${app.key}/billing`,
-          active: isActivePage('billing'),
-          allowed: allowedAccessTo(app, 'billing'),
-        },
-        // { id: 'Authentication', icon: <ShuffleIcon />, active: isActivePage("user_auto_messages")},
-      ],
+      children: app_settings_items(app, isActivePage),
     },
   ];
 
@@ -495,6 +427,7 @@ function Sidebar({
                       <Link
                         key={`sidebar-section-child-${id}-${childId}`}
                         to={url}
+                        aria-label={label}
                         disabled={!allowed}
                         className={`
                         ${active ? 'bg-gray-200 dark:bg-black' : ''}
