+++ config/permissions.yml
@@ -20,6 +20,7 @@ program_manager:
     - conversations
     - reports
     - segments
+    - team
   read:
     - users
   write:
