+++ spec/cypress/integration/app/agent_invitation.js
@@ -12,11 +12,11 @@ describe('Login Spec', function () {
       a.raw_invitation_token
     `).then((token) => {
       cy.visit(`/agents/invitation/accept?invitation_token=${token}`).then(() => {
-        cy.contains('Set password').should('exist')
+        cy.contains('Type your password').should('exist')
         cy.wait(1000)
         cy.xpath('//*[@id="password"]').type('123456')
         cy.xpath('//*[@id="password_confirmation"]').type('123456')
-        cy.contains('Set my password').click()
+        cy.contains('Save').click()
         cy.contains('my app').should('exist')
       })
     })
